refactor panel file and split some of common widgets inro separate files + modify the account tile & make pageWithDrawerOnDrag
